Visualize the DAG

Visualize the DAG

dag_circular_viz(
  dag,
  highlight = NULL,
  start = 0,
  end = 360,
  reorder_level = NA,
  partition_level = 0,
  node_col = NULL,
  node_transparency = 0.5,
  node_size = NULL,
  edge_col = NULL,
  edge_transparency = 0.98,
  legend_labels_from = NULL,
  legend_labels_max_width = 50
)

dag_as_DOT(
  dag,
  color = "black",
  style = "solid",
  fontcolor = "black",
  fontsize = 10,
  shape = "box",
  edge_color = NULL,
  edge_style = NULL
)

dag_graphviz(
  dag,
  color = "black",
  fontcolor = "black",
  fontsize = 10,
  shape = "box",
  edge_color = NULL,
  edge_style = NULL,
  ...
)

Arguments

dag

An ontology_Dag object.

highlight

A vector of terms to be highlighted on the DAG.

start

Start of the circle, measured in degree.

end

End of the circle, measured in degree.

reorder_level

Whether to reorder child terms. See dag_reorder().

partition_level

If node_col is not set, users can cut the DAG into clusters with different node colors.

node_col

Colors of nodes. If the value is a vector, the order should correspond to terms in dag_all_terms().

node_transparency

Transparency of nodes. The same format as node_col.

node_size

Size of nodes. The same format as node_col.

edge_col

A named vector where names correspond to relation types.

edge_transparency

A named vector where names correspond to relation types.

legend_labels_from

If partitioning is applied on the DAG, a legend is generated showing different top terms. By default, the legend labels are the term IDs. If there are additionally column stored in the meta data frame of the DAG object, the column name can be set here to replace the term IDs as legend labels.

legend_labels_max_width

Maximal width of legend labels. Labels are wrapped into multiple lines if the widths exceed it.

color

A vector of colors. If the value is a vector, the order should correspond to terms in dag_all_terms().

style

Style of the nodes. See https://graphviz.org/docs/attr-types/style/ for possible values.

fontcolor

Color of labels.

fontsize

Font size of labels.

shape

Shape of nodes. See https://graphviz.org/doc/info/shapes.html#polygon for possible values.

edge_color

A named vector where names correspond to relation types.

edge_style

A named vector where names correspond to relation types. See https://graphviz.org/docs/attr-types/style/ for possible values for edges.

...

Pass to DiagrammeR::grViz().

Details

dag_circular_viz() uses a circular layout for visualizing large DAGs. dag_graphviz() uses a hierarchical layout for visualizing small DAGs.

dag_as_DOT() generates the DOT code of the DAG.

dag_graphviz() visualizes the DAG with the DiagrammeR package.

See also